==> preprocess_notebook/0.3.mimic_extract.ipynb <==
import numpy as np
import pandas as pd
import os
import datetime
import sys
import tqdm


mimic_path = '/ssd1/mimic3'
save_path = '../data'
mimic_extract_folder = '/home/sanghoon/drug/EHR/MIMIC_Extract/'  # use your MIMIC_Extract folder
mimic_folder = os.getcwd()

## preprocess_ mimic_extract

mimic_path = '/ssd1/mimic3'
using_feature = ['ICUSTAYS', 'D_ITEMS','ADMISSIONS']
data_dict ={}
for file in using_feature:
    temp_name = file +'.csv'
    
    data_dict[file.lower()] = pd.read_csv(mimic_path+'/'+temp_name, low_memory=False)
    
chart_lab_icu = pd.read_pickle(save_path+'/chart_lab_icu.pkl')



os.chdir(mimic_extract_folder)

from datapackage_io_util import (
    load_datapackage_schema,
    load_sanitized_df_from_csv,
    save_sanitized_df_to_csv,
    sanitize_df,
)

os.chdir(mimic_folder)

def get_variable_mapping(mimic_mapping_filename):
    # Read in the second level mapping of the itemids
    var_map = pd.read_csv(mimic_mapping_filename, index_col=None)
    var_map = var_map.loc[(var_map['LEVEL2'] != '') & (var_map['COUNT']>0)]
    var_map = var_map.loc[(var_map['STATUS'] == 'ready')]
    var_map['ITEMID'] = var_map['ITEMID'].astype(int)

    return var_map

def get_variable_ranges(range_filename):
    # Read in the second level mapping of the itemid, and take those values out
    columns = [ 'LEVEL2', 'OUTLIER LOW', 'VALID LOW', 'IMPUTE', 'VALID HIGH', 'OUTLIER HIGH' ]
    to_rename = dict(zip(columns, [ c.replace(' ', '_') for c in columns ]))
    to_rename['LEVEL2'] = 'VARIABLE'
    var_ranges = pd.read_csv(range_filename, index_col=None)
    var_ranges = var_ranges[columns]
    var_ranges.rename(columns=to_rename, inplace=True)
    var_ranges = var_ranges.drop_duplicates(subset='VARIABLE', keep='first')
    var_ranges['VARIABLE'] = var_ranges['VARIABLE'].str.lower()
    var_ranges.set_index('VARIABLE', inplace=True)
    var_ranges = var_ranges.loc[var_ranges.notnull().all(axis=1)]

    return var_ranges



mimic_mapping_filename = os.path.join(mimic_extract_folder+'resources', 'itemid_to_variable_map.csv')
range_filename = os.path.join(mimic_extract_folder+'resources', 'variable_ranges.csv')

var_map = get_variable_mapping(mimic_mapping_filename)
var_ranges = get_variable_ranges(range_filename)

static_filename = 'static_data.csv'
static_columns_filename = 'static_colnames.txt'

dynamic_filename = 'vitals_hourly_data.csv'
columns_filename = 'vitals_colnames.txt'
subjects_filename = 'subjects.npy'
times_filename = 'fenceposts.npy'
dynamic_hd5_filename = 'vitals_hourly_data.h5'
dynamic_hd5_filt_filename = 'all_hourly_data.h5'

codes_hd5_filename = 'C.h5'
notes_hd5_filename = 'notes.hdf' # N.h5
idx_hd5_filename = 'C_idx.h5'

outcome_filename = 'outcomes_hourly_data.csv'
outcome_hd5_filename = 'outcomes_hourly_data.h5'
outcome_columns_filename = 'outcomes_colnames.txt'

# SQL command params

ID_COLS = ['subject_id', 'hadm_id', 'icustay_id']
ITEM_COLS = ['itemid', 'label', 'LEVEL1', 'LEVEL2']

args={'resource_path':mimic_extract_folder+'resources',
     'out_path':save_path,
     'pop_size':0,
     'extract_pop':1,
      'min_age':15,
      'min_duration':12,
      'max_duration':2400
     }

def apply_variable_limits(df, var_ranges, var_names_index_col='LEVEL2'):
    idx_vals        = df.index.get_level_values(var_names_index_col)
    non_null_idx    = ~df.value.isnull()
    var_names       = set(idx_vals)
    var_range_names = set(var_ranges.index.values)

    for var_name in tqdm.tqdm(var_names):
        var_name_lower = var_name.lower()
        if var_name_lower not in var_range_names:
            print("No known ranges for %s" % var_name)
            continue

        outlier_low_val, outlier_high_val, valid_low_val, valid_high_val = [
            var_ranges.loc[var_name_lower, x] for x in ('OUTLIER_LOW','OUTLIER_HIGH','VALID_LOW','VALID_HIGH')
        ]

        running_idx = non_null_idx & (idx_vals == var_name)

        outlier_low_idx  = (df.value < outlier_low_val)
        outlier_high_idx = (df.value > outlier_high_val)
        valid_low_idx    = ~outlier_low_idx & (df.value < valid_low_val)
        valid_high_idx   = ~outlier_high_idx & (df.value > valid_high_val)

        var_outlier_idx   = running_idx & (outlier_low_idx | outlier_high_idx)
        var_valid_low_idx = running_idx & valid_low_idx
        var_valid_high_idx = running_idx & valid_high_idx

        df.loc[var_outlier_idx, 'value'] = np.nan
        df.loc[var_valid_low_idx, 'value'] = valid_low_val
        df.loc[var_valid_high_idx, 'value'] = valid_high_val

        n_outlier = sum(var_outlier_idx)
        n_valid_low = sum(var_valid_low_idx)
        n_valid_high = sum(var_valid_high_idx)
        if n_outlier + n_valid_low + n_valid_high > 0:
            print(
                "%s had %d / %d rows cleaned:\n"
                "  %d rows were strict outliers, set to np.nan\n"
                "  %d rows were low valid outliers, set to %.2f\n"
                "  %d rows were high valid outliers, set to %.2f\n"
                "" % (
                    var_name,
                    n_outlier + n_valid_low + n_valid_high, sum(running_idx),
                    n_outlier, n_valid_low, valid_low_val, n_valid_high, valid_high_val
                )
            )

    return df

UNIT_CONVERSIONS = [
    ('weight',                   'oz',  None,             lambda x: x/16.*0.45359237),
    ('weight',                   'lbs', None,             lambda x: x*0.45359237),
    ('fraction inspired oxygen', None,  lambda x: x > 1,  lambda x: x/100.),
    ('oxygen saturation',        None,  lambda x: x <= 1, lambda x: x*100.),
    ('temperature',              'f',   lambda x: x > 79, lambda x: (x - 32) * 5./9),
    ('height',                   'in',  None,             lambda x: x*2.54),
]
def get_values_by_name_from_df_column_or_index(data_df, colname):
    """ Easily get values for named field, whether a column or an index

    Returns
    -------
    values : 1D array
    """
    try:
        values = data_df[colname]
    except KeyError as e:
        if colname in data_df.index.names:
            values = data_df.index.get_level_values(colname)
        else:
            raise e
    return values

def standardize_units(X, name_col='itemid', unit_col='valueuom', value_col='value', inplace=True):
    if not inplace: X = X.copy()
    name_col_vals = get_values_by_name_from_df_column_or_index(X, name_col)
    unit_col_vals = get_values_by_name_from_df_column_or_index(X, unit_col)

    try:
        name_col_vals = name_col_vals.str
        unit_col_vals = unit_col_vals.str
    except:
        print("Can't call *.str")
        print(name_col_vals)
        print(unit_col_vals)
        raise

    #name_filter, unit_filter = [
    #    (lambda n: col.contains(n, case=False, na=False)) for col in (name_col_vals, unit_col_vals)
    #]
    # TODO(mmd): Why does the above not work, but the below does?
    name_filter = lambda n: name_col_vals.contains(n, case=False, na=False)
    unit_filter = lambda n: unit_col_vals.contains(n, case=False, na=False)

    for name, unit, rng_check_fn, convert_fn in tqdm.tqdm(UNIT_CONVERSIONS):
        name_filter_idx = name_filter(name)
        needs_conversion_filter_idx = name_filter_idx & False

        if unit is not None: needs_conversion_filter_idx |= name_filter(unit) | unit_filter(unit)
        if rng_check_fn is not None: needs_conversion_filter_idx |= rng_check_fn(X[value_col])

        idx = name_filter_idx & needs_conversion_filter_idx

        X.loc[idx, value_col] = convert_fn(X[value_col][idx])

    return X

args['out_path']

static_data_schema = load_datapackage_schema(
        os.path.join(args['resource_path'], 'static_data_spec.json'))
outcome_data_schema = load_datapackage_schema(
    os.path.join(args['resource_path'], 'outcome_data_spec.json'))

if not os.path.isdir(args['out_path']):
    print('ERROR: OUTPATH %s DOES NOT EXIST' % args['out_path'])
    sys.exit()
else:
    outPath = args['out_path']


# Modify the filenames
if args['pop_size'] > 0:
    pop_size = str(args['pop_size'])

    static_filename = splitext(static_filename)[0] + '_' + pop_size + splitext(static_filename)[1]
    dynamic_filename = splitext(dynamic_filename)[0] + '_' + pop_size + splitext(dynamic_filename)[1]
    #columns_filename = splitext(columns_filename)[0] + '_' + pop_size + splitext(columns_filename)[1]
    subjects_filename = splitext(subjects_filename)[0] + '_' + pop_size + splitext(subjects_filename)[1]
    times_filename = splitext(times_filename)[0] + '_' + pop_size + splitext(times_filename)[1]
    dynamic_hd5_filename = splitext(dynamic_hd5_filename)[0] + '_' + pop_size + splitext(dynamic_hd5_filename)[1]
    outcome_filename = splitext(outcome_filename)[0] + '_' + pop_size + splitext(outcome_filename)[1]
    dynamic_hd5_filt_filename = splitext(dynamic_hd5_filt_filename)[0] + '_' + pop_size + splitext(dynamic_hd5_filt_filename)[1]
    outcome_hd5_filename = splitext(outcome_hd5_filename)[0] + '_' + pop_size + splitext(outcome_hd5_filename)[1]
    #outcome_columns_filename = splitext(outcome_columns_filename)[0] + '_' + pop_size + splitext(outcome_columns_filename)[1]
    codes_hd5_filename = splitext(codes_hd5_filename)[0] + '_' + pop_size + splitext(codes_hd5_filename)[1]
    notes_hd5_filename = splitext(notes_hd5_filename)[0] + '_' + pop_size + splitext(notes_hd5_filename)[1]
    idx_hd5_filename = splitext(idx_hd5_filename)[0] + '_' + pop_size + splitext(idx_hd5_filename)[1]



#############
# Population extraction

data = None
if (args['extract_pop'] == 0 | (args['extract_pop'] == 1) ) & os.path.isfile(os.path.join(outPath, static_filename)):
    print("Reloading data from %s" % os.path.join(outPath, static_filename))
    data = pd.read_csv(os.path.join(outPath, static_filename))
    data = sanitize_df(data, static_data_schema)
elif (args['extract_pop'] == 1 & (not os.path.isfile(os.path.join(outPath, static_filename)))) | (args['extract_pop'] == 2):
    print("Building data from scratch.")
    pop_size_string = ''
    if args['pop_size'] > 0:
        pop_size_string = 'LIMIT ' + str(args['pop_size'])

    min_age_string = str(args['min_age'])
    min_dur_string = str(args['min_duration'])
    max_dur_string = str(args['max_duration'])
    min_day_string = str(float(args['min_duration'])/24)

    template_vars = dict(
        limit=pop_size_string, min_age=min_age_string, min_dur=min_dur_string, max_dur=max_dur_string,
        min_day=min_day_string
    )

var_map = var_map[['LEVEL2', 'ITEMID', 'LEVEL1']].rename({'LEVEL2': 'LEVEL2', 'LEVEL1': 'LEVEL1', 'ITEMID': 'itemid'}, axis=1).set_index('itemid')
chart_lab_icu.columns = list(map(lambda x :x.lower(), chart_lab_icu.columns))
chart_lab_icu['value']= pd.to_numeric(chart_lab_icu['value'],'coerce')
chart_lab_icu.astype({k: int for k in ID_COLS })

to_hours = lambda x: max(0, x.days*24 + x.seconds // 3600)

chart_lab_icu = chart_lab_icu.set_index('icustay_id')
chart_lab_icu['hours_in'] = (chart_lab_icu['charttime'] - chart_lab_icu['intime']).apply(to_hours)

chart_lab_icu.drop(columns=['charttime', 'intime'], inplace=True)
chart_lab_icu.set_index('itemid', append=True, inplace=True)


items_d= data_dict['d_items']
items_d.columns = list(map(lambda x :x.lower(), items_d.columns))
items_d=items_d[['itemid','label','dbsource','linksto','category','unitname']]



items_d=items_d.set_index('itemid')
chart_lab_icu = chart_lab_icu.join(var_map).join(items_d).set_index(['label', 'LEVEL1', 'LEVEL2'], append=True)


standardize_units(chart_lab_icu, name_col='LEVEL1', inplace=True)

chart_lab_icu_drop_level2na = chart_lab_icu[~chart_lab_icu.index.get_level_values('LEVEL2').isna()]

chart_lab_icu_lmits = apply_variable_limits(chart_lab_icu_drop_level2na, var_ranges, 'LEVEL2')

t=chart_lab_icu_lmits[chart_lab_icu_lmits['hours_in']<=23] # check only 24 hours in lab items
t=t[['subject_id','hadm_id','value','hours_in']]

group_item_cols = ['LEVEL2']
t_groups = t.groupby(ID_COLS+group_item_cols+['hours_in']).agg(['mean','std','count'])
t_groups.columns = t_groups.columns.droplevel(0)
t_groups.columns.names = ['Aggregation Function']
t_groups=t_groups.unstack(level='LEVEL2')

t_groups.columns = t_groups.columns.reorder_levels(order=group_item_cols + ['Aggregation Function'])

t_groups=t_groups.sort_index(axis=0).sort_index(axis=1)
idx = pd.IndexSlice
t_groups.loc[:, idx[:, 'count']] = t_groups.loc[:, idx[:, 'count']].fillna(0)


print("Shape of X : ", t_groups.shape) # in 24 hours


temp_check= t_groups.reset_index()

len(temp_check['subject_id'].unique())

t_groups.to_hdf(save_path+'/vitals_24hourly_data_sh.h5','X')

# make label

del data_dict['icustays']['ROW_ID']
del data_dict['admissions']['ROW_ID']
concat_label=data_dict['icustays'].merge(data_dict['admissions'], 'left', on=['HADM_ID','SUBJECT_ID'])
concat_label['INTIME'] = pd.to_datetime(concat_label['INTIME'])
concat_label['OUTTIME'] = pd.to_datetime(concat_label['OUTTIME'])
concat_label['ADMITTIME'] = pd.to_datetime(concat_label['ADMITTIME'])
concat_label['DISCHTIME'] = pd.to_datetime(concat_label['DISCHTIME'])
concat_label['DEATHTIME'] = pd.to_datetime(concat_label['DEATHTIME'])

concat_label['mort_icu']= list(map(lambda x : 1 if x[0]<=x[2]<=x[1] else 0,concat_label[['INTIME','OUTTIME','DEATHTIME']].values))
concat_label['mort_hosp']= list(map(lambda x : 1 if x[0]<=x[2]<=x[1] else 0,concat_label[['ADMITTIME','DISCHTIME','DEATHTIME']].values))

concat_label.columns = list(map(lambda x : x.lower(), concat_label.columns))
concat_label['max_hours']=(concat_label['outtime']-concat_label['intime']).apply(to_hours)

idx_have_ts=temp_check.icustay_id.unique()

concat_label_ts=concat_label[concat_label['icustay_id'].isin(idx_have_ts)]

concat_label_ts[['subject_id','hadm_id','icustay_id','max_hours','mort_icu','mort_hosp','los']].to_csv(save_path + '/label_have_ts_los.csv')

# make note

DATAPATH = mimic_path 
noteevents_df = pd.read_csv(os.path.join(DATAPATH, "NOTEEVENTS.csv"))


noteevents_df.groupby(noteevents_df.CATEGORY).agg(['count'])
note_categories = noteevents_df.groupby(noteevents_df.CATEGORY).agg(['count']).index
selected_note_types = []
for each_cat in tqdm.tqdm(list(note_categories)):
    if each_cat != 'Discharge summary':
        selected_note_types.append(each_cat)
        
sub_notes = noteevents_df[noteevents_df.CATEGORY.isin(selected_note_types)]
sub_notes.shape

missing_chardate_index = []
for each_note in tqdm.tqdm(sub_notes.itertuples()):
    if isinstance(each_note.CHARTTIME, str):
        continue
    if np.isnan(each_note.CHARTTIME):
        missing_chardate_index.append(each_note.Index)
print ("{} of notes does not charttime.".format(len(missing_chardate_index)))
print(sub_notes.shape)

sub_notes.drop(missing_chardate_index, inplace=True )
print(sub_notes.shape)

sub_notes_have_ts=sub_notes[sub_notes.SUBJECT_ID.isin(concat_label_ts.subject_id.unique())]
sub_notes_have_ts.shape , len(sub_notes_have_ts.SUBJECT_ID.unique())

have_notes_ids=sub_notes_have_ts.SUBJECT_ID.unique()

sub_notes_have_ts.to_csv(save_path+'/sub_note.csv',index=False)
