==> preprocess_notebook/2.make_note_to_feature.ipynb <==
import pandas as pd
import os
import numpy as np
import tqdm
import warnings
warnings.filterwarnings('ignore')
import ast

def check_max_len_text(x):
    max_len=-1
    max_len_text =''
    for time_doc in x:
        
        temp_doc = ' '.join(time_doc)
        if len(temp_doc)>max_len:
            max_len = len(temp_doc)
            max_len_text = temp_doc
    
    return max_len_text

def check_max_len_sentence(x):
    max_len=-1
    max_len_text =''
    for time_doc in x:
        
        temp_doc = ' '.join(time_doc)
        if len(temp_doc)>max_len:
            max_len = len(temp_doc)
            max_len_text = time_doc
    
    return max_len_text

data_path = '../data/'
preprocessed_notes = pd.read_pickle(data_path+'preprocessed_notes.pkl')
labels=pd.read_csv(data_path+'label_last_visit.csv')
split_path = data_path + 'split_0/'
train_subj = pd.read_pickle(split_path+'train_hadm_idx.pkl')
dev_subj = pd.read_pickle(split_path+'dev_hadm_idx.pkl')
test_subj = pd.read_pickle(split_path+'test_hadm_idx.pkl')


len(train_subj), len(dev_subj), len(test_subj)

preprocessed_notes=preprocessed_notes.sort_values('CHARTTIME')
#preprocessed_notes['preprocessed_text']= preprocessed_notes['preprocessed_text'].apply(lambda x: ast.literal_eval(x))

# for bert tokeb

max_len_text=preprocessed_notes.groupby(['subject_id','hadm_id_y','CHARTTIME']).preprocessed_text.apply(check_max_len_sentence)
max_len_text=max_len_text.reset_index()


max_len_text['preprocessed_text'][0]

tag=preprocessed_notes[['hadm_id_y','CHARTTIME','CATEGORY','DESCRIPTION']]
tag=tag[tag.duplicated()==False]


max_len_text=max_len_text.merge(tag,'left',on=['hadm_id_y','CHARTTIME'])

max_len_text.groupby('hadm_id_y').count().max() # 27 document.

max_len_text['count_sentence'] = max_len_text['preprocessed_text'].apply(lambda x : len(x))

def check_max_token(x):
    max_len=0
    for text in x:
        temp_len=len(text.split(' '))
        if temp_len > max_len:
            max_len=temp_len
    return max_len

max_len_text['max_token'] = max_len_text['preprocessed_text'].apply(lambda x : len(x))

max_len_text.groupby('CATEGORY').count()

max_len_text_r=max_len_text[max_len_text['CATEGORY'].isin(['Social Work', 'Case Management ', 'Rehab Services' , 'Nutrition'])==False]

max_len_text_r= max_len_text_r[max_len_text_r['max_token']!=0]

max_len_text_r.groupby('hadm_id_y').count_sentence.sum().max()

max_len_text_r.to_pickle(data_path+'for_bert_text.pkl')

# run ../src/make_emb.py to make sentence_to_cbert_cls.npy

# make bert emb shape

text= pd.read_pickle(data_path+'for_bert_text.pkl')

cbert_emb=np.load(data_path+'sentence_to_cbert_cls.npy')

group_st_c = text.groupby('hadm_id_y').count_sentence.sum()

(group_st_c>512).sum()

start=0
end=0
idx_start=[]
for i in tqdm.tqdm(range(len(text))):
    temp_c=text.iloc[i]['count_sentence']
    idx_start.append(start)
    end = start+temp_c
    start = end
    

text['idx_start']=idx_start

cbert_dict_emb={}
hadm_ids=text['hadm_id_y'].unique()
for key in tqdm.tqdm(hadm_ids):
    temp_notes=text[text['hadm_id_y']==key]
    temp_numpy_list=[]
    for i in range(len(temp_notes)):
        start , c_s =temp_notes.iloc[i][['idx_start','count_sentence']].values
        temp_emb=cbert_emb[start:start+c_s]
        temp_numpy_list.append(temp_emb)
    cbert_dict_emb[key] = np.concatenate(temp_numpy_list)

pd.to_pickle(cbert_dict_emb,data_path+'cbert_emb_by_hadm.pkl')
