==> preprocess_notebook/3.ccs_labeling.ipynb <==
import pandas as pd
import json
import os
import numpy as np
import ast
import tqdm

data_path = '/ssd2/sanghoon/data/EHR/mimic3/mimic_extract/sh_pre_los100/24h_6g/'
mimic_path='/ssd2/sanghoon/data/EHR/mimic3/mimic3/'
patitentTM_preprocessing_path='/home/sanghoon/drug/project/EHR/PatientTM/data/extended/preprocessing/'
labels=pd.read_csv(data_path+'label_last_visit.csv')
max_len_text_r =pd.read_pickle(data_path+'for_bert_text.pkl')

drg_code = pd.read_csv(mimic_path+'DRGCODES.csv')
ccs_code = pd.read_json(patitentTM_preprocessing_path+'ICDandCCSmappings/merged_icdccs_codes.json', typ='series')
ccs_code=ccs_code.reset_index()
diag=pd.read_csv(mimic_path+'DIAGNOSES_ICD.csv')

diag=diag[diag['HADM_ID'].isin(labels['hadm_id'].unique())]
pre_data=pd.read_csv('/ssd2/sanghoon/data/EHR/MP_EHR/preprocessed.csv')
labels=labels[labels['have_note']==1]
pre_data=pre_data[pre_data['HADM_ID'].isin(labels['hadm_id'])]


all_diag_ccs = pre_data['DIAG_CCS'].apply(lambda x: ast.literal_eval(x))

all_d_label = list(set(np.concatenate(all_diag_ccs.values)))

d_ccs_label_dict=dict(zip(all_d_label, [i for i in range(len(all_d_label))]))
len(d_ccs_label_dict)

def  make_label_idx_list(t):
    
    return list(map(lambda x: d_ccs_label_dict[x],t))

label_dict={}
for i in tqdm.tqdm(range(len(pre_data))):
    key = pre_data.iloc[i]['HADM_ID']
    label_t = list(map(lambda x : d_ccs_label_dict[x] ,all_diag_ccs.iloc[i]))
    temp_label=np.zeros([1,len(d_ccs_label_dict)])
    temp_label[0][label_t]=1
    label_dict[key]=temp_label
    

pd.to_pickle(label_dict, data_path+'/ccs_label.pkl')
